+++ test/models/quantity_test.rb
@@ -1,0 +1,7 @@
+require "test_helper"
+
+class QuantityTest < ActiveSupport::TestCase
+  # test "the truth" do
+  #   assert true
+  # end
+end
